+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 Unreleased
 ----------
-**Changed**
+**Improvements**
  - public_model task also defines methods mapped to MAS module steps when publishing to MAS.
+ - SSL verification can be disable with `SSLREQCERT` environment variable.
 
 
 v0.9.6 (2019-07-15)